refactor useEffect for Achievement notifications into custom hook
