+++ src/Screens/Achievements/index.tsx
@@ -1,7 +1,5 @@
 import { View, Button } from 'react-native';
-import notifee, { AuthorizationStatus } from '@notifee/react-native';
 import React from 'react';
-import { useRef, useEffect, useState } from 'react';
 import BackgroundWrapper from '../../Components/BackgroundWrapper';
 import Text from '../../Components/Text';
 import styles from './defaultCSS';
@@ -9,7 +7,7 @@ import Trophy from '../../Assets/svg/Trophy.svg';
 import Award from '../../Assets/svg/Award.svg';
 import { Appbar } from 'react-native-paper';
 import useAchievements from './data';
-import { useNotification } from '../../Features/useNotification';
+import { useAchievementNotifications } from '../../Hooks/useAchievementNotifications';
 
 
 interface Achievement {
@@ -21,48 +19,20 @@ interface Achievement {
 }
 
 const AchievementsList = ({ achievements }: { achievements: Achievement[] }) => (
-  // console.log(achievements),  // Check if achievements are being passed;
-
   <View>
     {achievements.map((item) => (
       <View key={item.id} >
         <Text style={styles.itemText}>• {item.points} {item.description}</Text>
       </View>
     ))}
-
   </View>
 );
 
 const Index: React.FC<{ navigation: any }> = ({ navigation }) => {
   const user_id = 1; //replace with the actual user id of the logged in user
   const { achievements, loading, error } = useAchievements(user_id);
-  const { displayNotification } = useNotification();
-  const displayedNotifications = useRef<Set<number>>(new Set());
 
-  useEffect(() => {
-    const displayNotifications = async () => {
-      if (achievements) {
-        for (const achievement of achievements) {
-          if (achievement.points > 1 && !displayedNotifications.current.has(achievement.id)) { // Replace 1 with your condition
-            console.log(`Displaying notification for: ${achievement.description} with ${achievement.points} points`);
-            await displayNotification('Achievement Unlocked', `${achievement.points} ${achievement.description}`);
-            displayedNotifications.current.add(achievement.id);
-            await new Promise(resolve => setTimeout(resolve, 1000)); // Add a 1-second delay
-          }
-        }
-
-      }
-    };
-    displayNotifications();
-  }, [achievements, displayNotification]);
-
-
-  if (loading) {
-    return <Text>Loading...</Text>;
-  }
-  // const handleNotification = (title, body) => {
-  //   displayNotification(title, body);
-  // };
+  useAchievementNotifications(achievements);
 
   if (loading) {
     return <Text>Loading...</Text>;
